eventbus/requirements.txt
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements --dev
#

# Note: in pipenv 2020.x, "--dev" changed to emit both default and development
# requirements. To emit only development requirements, pass "--dev-only".

-i https://pypi.org/simple
flask-cors==3.0.10
flask==1.1.2
mongoengine==0.23.1
mongomock==3.22.1
# pyjwt==2.0.1
pylint==2.7.2
pymongo==3.11.4
pytest==6.2.2
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.25.1
Flask-Session==0.3.2
Flask-Ext==0.1
regex==2021.4.4
mongoengine==0.23.1
# bcrypt==3.2.0
flask-mongoengine==1.0.0
# cryptography==3.4.7
# Flask-SQLAlchemy==2.5.1
stripe==2.57.0
requests==2.25.1
DateTime==4.3
multiprocess==0.70.11
asyncio==3.4.3